third_party/blink/web_tests/platform/mac-mac14/external/wpt/css/css-scrollbars/scrollbar-color-003-expected.txt

This is a testharness.js-based test.
[FAIL] scrollbar-color yellow blue overrides ::-webkit-scrollbar
  assert_less_than: themed scrollWidth expected a number less than 200 but got 200
Harness: the test ran to completion.
